updating student repo

+++ Domain/Entities/Scholarship.cs
@@ -15,7 +15,7 @@ namespace Domain.Entities
         public ScholarshipType Type { get; set; } // Enum: Amount or Percentage
 
         public decimal? Amount { get; set; }       // Only if Type == Amount
-        public float? Percentage { get; set; }     // Only if Type == Percentage
+        public decimal? Percentage { get; set; }     // Only if Type == Percentage
 
         public string? Description { get; set; }   // Optional notes (e.g. scholarship name)
